menuiserie

+++ database/migrations/2014_10_12_000000_create_quotes_table.php
@@ -23,8 +23,8 @@ class CreateQuotesTable extends Migration
             $table->date('date')->nullable();
             $table->tinyInteger('menuiserie')->nullable();
 
-            $table->integer('tva')->default(0)->nullable();
-            $table->integer('remise')->default(0)->nullable();
+            $table->double('tva', 8, 2)->default(0)->nullable();
+            $table->double('remise', 8, 2)->default(0)->nullable();
 
             $table->double('total_tva', 8, 2)->default(0)->nullable();
             $table->double('total_remise', 8, 2)->default(0)->nullable();
