+++ database/migrations/2014_10_14_000000_create_items_table.php
@@ -17,17 +17,17 @@ class CreateItemsTable extends Migration
             $table->bigIncrements('id');
             $table->text('description')->nullable();
             $table->integer('qty')->default(1);
-            $table->integer('tva');
+            $table->double('tva', 8, 2);
 
             $table->unsignedBigInteger('product')->nullable();
             $table->foreign('product')->references('id')->on('products')->onDelete('cascade');
-            
+
 			$table->unsignedBigInteger('quote')->nullable();
             $table->foreign('quote')->references('id')->on('quotes')->onDelete('cascade');
-			
+
 			$table->unsignedBigInteger('invoice')->nullable();
             $table->foreign('invoice')->references('id')->on('invoices')->onDelete('cascade');
-			
+
             $table->timestamps();
         });
     }
